♻️ refactor: extract model detail panel hook
