+++ src/features/ModelSwitchPanel/types.ts
@@ -4,9 +4,9 @@ import { type ComponentType } from 'react';
 
 import { type EnabledProviderWithModels } from '@/types/aiProvider';
 
-import { type PricingMode } from './components/ModelDetailPanel';
-
 export type GroupMode = 'byModel' | 'byProvider';
+
+export type PricingMode = 'image' | 'video';
 
 export interface ModelWithProviders {
   displayName: string;
